moved ScoreContext.tsx to components
